Process linker script before doing sanity checks
